added checking hash

+++ memes/models.py
@@ -15,7 +15,8 @@ from searchEngine.indexer import misc
 from searchEngine.indexer import simplifier
 from tags.models import Tags
 
-#import memes.algo.img2hash as img2hash
+import memes.algo.img2hash as img2hash
+from PIL import Image
 
 
 def update_index_in_db(text, description, new_index_text, new_index_description):
@@ -103,40 +104,60 @@ class Memes(models.Model):
     def delete(self, **kwargs):
         pass
 
+    def _check_img_uniqueness(self):
+        if self.image is not None and self.image != '':
+            self.image_hash = img2hash.hash_from_image(Image.open(self.image))
+
+            # проверяем есть ли у нас мем с таким хешом, и если нет, то бросается исключение, которое игнорим.
+            # если исключение не бросится, значит мем с таким хешом существует и мы не сохраняем что сейчас имеем
+            try:
+                ext_meme = Memes.objects.get(Q(image_hash=self.image_hash))
+                print("МЕМ НЕ БУДЕТ ДОБАВЛЕН")
+                # TODO: union memes
+                return False
+            except Memes.DoesNotExist:
+                print("МЕМ БУДЕТ ДОБАВЛЕН")
+                return True  # meme is uniq
+
     def save(self, *args, **kwargs):
-        first_save = False
-        if self.image is not None and self.image != '':
-            first_save = True
-            # Разметка мема (текст)
-            # self.textDescription = re.sub("[^а-яa-z0-9 ]+", "", " ".join(getTextFromImage(self.image)).lower())
+        is_uniq_img = True  # self._check_img_uniqueness()  # если true, то мем будет добавлен
 
-            # Сохранение мема на яндекс.диск
-            y = settings.Y
-            name = self.image.url.split(".")
-            self.time = time.time()
-            self.fileName = ".".join(name[:-1]) + "_{}".format(self.time) + ".jpg"
+        if is_uniq_img:  # self.image_hash мы ставим в функции проверки _check_img_uniq...
+            first_save = False
+            if self.image is not None and self.image != '':
+                first_save = True
+                # Разметка мема (текст)
+                # self.textDescription = re.sub("[^а-яa-z0-9 ]+", "", " ".join(getTextFromImage(self.image)).lower())
 
-        self.rating = self.likes - self.dislikes
-        if self.dislikes != 0:
-            self.ratio = self.likes / self.dislikes
-        else:
-            self.ration = self.likes
-        super(Memes, self).save(*args, **kwargs)
+                # Сохранение мема на яндекс.диск
+                y = settings.Y
+                name = self.image.url.split(".")
+                self.time = time.time()
+                self.fileName = ".".join(name[:-1]) + "_{}".format(self.time) + ".jpg"
 
-        # Построение нового индекса по добавленному мему
-        meme_index = indexer.full_index([info.MemeInfo(self.id, self.textDescription, self.imageDescription)])
-        update_index_in_db(self.textDescription, self.imageDescription, meme_index.text_words, meme_index.descr_words)
+            self.rating = self.likes - self.dislikes
+            if self.dislikes != 0:
+                self.ratio = self.likes / self.dislikes
+            else:
+                self.ration = self.likes
+            super(Memes, self).save(*args, **kwargs)
 
-        if self.image is not None and self.image != '':
-            if os.path.isfile(self.image.path):
-                # сжатие картинки
-                compress_image(self.image.path, self.fileName[1:])
-                y.upload(self.fileName[1:], self.fileName)
-                self.url = yadisk.functions.resources.get_download_link(y.get_session(), self.fileName)
-                os.remove(self.image.path)
-                os.remove(self.fileName[1:])
-                self.image = None
-                super(Memes, self).save(update_fields=['image', 'url'])
+            # Построение нового индекса по добавленному мему
+            meme_index = indexer.full_index([info.MemeInfo(self.id, self.textDescription, self.imageDescription)])
+            update_index_in_db(self.textDescription, self.imageDescription, meme_index.text_words, meme_index.descr_words)
 
-        if first_save and self.id % 100 == 0:
-            y.upload("db.sqlite3", "backup/db_{}.sqlite3".format(self.id))
+            if self.image is not None and self.image != '':
+                if os.path.isfile(self.image.path):
+                    # сжатие картинки
+                    compress_image(self.image.path, self.fileName[1:])
+                    y.upload(self.fileName[1:], self.fileName)
+                    self.url = yadisk.functions.resources.get_download_link(y.get_session(), self.fileName)
+                    os.remove(self.image.path)
+                    os.remove(self.fileName[1:])
+                    self.image = None
+                    super(Memes, self).save(update_fields=['image', 'url'])
+
+            if first_save and self.id % 100 == 0:
+                y.upload("db.sqlite3", "backup/db_{}.sqlite3".format(self.id))
+        else:  # такая картинка у нас уже есть
+            pass
